VMM/CPUM,HMVMX,IEM: Use the individual CPUMCTX_INHIBIT_SHADOW_SS and CPUMCTX_INHIBIT_SHADOW_STI flags with IEM (only setting), VT-x and NEM/kvm. [gcc build fix, 2nd try] bugref:9941
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@97186 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/vmm/cpum.h
@@ -2123,8 +2123,8 @@ DECLINLINE(void) CPUMUpdateInterruptShadowSsStiEx(PCPUMCTX pCtx, bool fInhibited
         pCtx->fInhibit &= (uint8_t)~CPUMCTX_INHIBIT_SHADOW;
     else
     {
-        pCtx->fInhibit |= (fInhibitedBySs  ? CPUMCTX_INHIBIT_SHADOW_SS  : (uint8_t)0)
-                       |  (fInhibitedBySti ? CPUMCTX_INHIBIT_SHADOW_STI : (uint8_t)0);
+        pCtx->fInhibit |= (fInhibitedBySs  ? (uint8_t)CPUMCTX_INHIBIT_SHADOW_SS  : (uint8_t)0)
+                       |  (fInhibitedBySti ? (uint8_t)CPUMCTX_INHIBIT_SHADOW_STI : (uint8_t)0);
         pCtx->uRipInhibitInt = rip;
     }
 }
